pyproject.toml
#
# Licensed to the Apache Software Foundation (ASF) under one
# or more contributor license agreements.  See the NOTICE file
# distributed with this work for additional information
# regarding copyright ownership.  The ASF licenses this file
# to you under the Apache License, Version 2.0 (the
# "License"); you may not use this file except in compliance
# with the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "airavata-cerebrum"
version = "0.1"
description = "Apache Airavata Cerebrum : Integrated Data-driver Modeling for Computational Neuroscience"
authors = [
    "Sriram P Chockalingam <srirampc@gatech.edu>"
]
maintainers = [
    "Apache Airavata Developers <dev@apache.airavata.org>"
]
readme = "README.md"
homepage = "https://github.com/apache/airavata-cerebrum"
license =  "Apache-2.0" 
keywords = [ "Data-driven", "Computational Neuroscience", "Neuroscience Modeling"]

[tool.poetry.dependencies]
python = "^3.10"
abc_atlas_access =  { git = "https://github.com/alleninstitute/abc_atlas_access" }
neuroanalysis = { git = "https://github.com/alleninstitute/neuroanalysis" }
aisynphys = { git = "https://github.com/alleninstitute/aisynphys" }
allensdk = "^2.16"
anndata = "^0.10"
airavata-python-sdk = "^1.1.6"
bmtk = "^1.1"
intervaltree = "^3.1.0"
ipywidgets = "^8.1"
jsonpath-python = "^1.2.0"
matplotlib = "^3.7"
numpy = "^1.23"
openpyxl = "^3.1"
pandas = "^1.5"
pydantic = "^2.7"
requests = "^2.31"
scikit-learn = "^1.5"
scanpy = "^1.9"
scipy = "^1.9"
tqdm = "^4.66"
pillow = "^10.4"
pyyaml = "^6.0"
statsmodels = "^0.14"

[tool.setuptools.packages.find]
include = ["airavata_cerebrum"]
